+++ entities/trade_test.go
@@ -103,7 +103,7 @@ func v2StylePool(token0, token1 *entities.Token, reserve0, reserve1 *entities.Cu
 	if err != nil {
 		panic(err)
 	}
-	pool := NewPoolV3(uint16(feeAmount), int32(tick), sqrtRatioX96, token0, token1, p)
+	pool := NewPoolV3(uint16(feeAmount), tick, sqrtRatioX96, token0, token1, p)
 	pool.Liquidity = uint256.MustFromBig(liquidity)
 	// if err != nil {
 	// 	panic(err)
